CheckError: Show single trader address in a chunk
Previously we would only show the trader address if it is a single
trader in the whole set.  We could show it if it is a single trader in
the chunk, as chunk is our execution unit.

Also updated the `to N` part of the message.  It was setting `N` to be
one above the actual index of the last trader in the chunk that we
checked.

Slightly simplify the `CheckError` interface - we only need the first
trader index and the total number of traders.

Also, avoid string to number conversion in `filterLiquidatableTraders`.